support .aoe2record files, fixes #9
This skips the aoe2record file header. I'm not sure yet what's in
there, except for some map and AI related names. It's very
variable length so the skipping is a bit haxorz.

There's probably a bunch of version information in there about
expansion packs or mods, so it'd be really good to be able to read
that someday 🙇